Updated terminology to be more generic.
