+++ docroot/modules/custom/uiowa_core/src/EntityProcessorInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Drupal\uiowa_core;
+
+/**
+ * Entity sync operation.
+ */
+interface EntityProcessorInterface {
+
+  /**
+   * Run the process operation.
+   */
+  public function process();
+
+}
