--- docroot/modules/custom/uiowa_core/src/EntitySyncInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Drupal\uiowa_core;
-
-/**
- * Entity sync operation.
- */
-interface EntitySyncInterface {
-
-  /**
-   * Run the sync operation.
-   */
-  public function sync();
-
-}
